app/requirements.txt
flask==2.2.5
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-flask==0.46b0
prometheus-client==0.20.0
prometheus_flask_exporter==0.22.4
